requirements.txt
discord.py
asyncio
pynacl
os